pyproject.toml
[project]
name = "doctomood"
version = "0.0.1"
description = "CLI tool to parse DOCX multiple-choice questions into Moodle XML and DOCX."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Ivan Alsina-Ferrer", email = "example@email.com" },
]
license = { text = "MIT" }

dependencies = [
  "ipython",
  "jupyter",
  "jupyterlab",
  "python-docx",
  "odfpy",
  "pandas",
  "pyyaml",
  "tqdm",
]

[project.optional-dependencies]
dev = [
  "black>=23.7.0",
  "isort>=5.12.0",
  "pylint>=3.0.0",
  "pre-commit>=4.5.1",
]

[project.scripts]
doctomood = "doctomood.main:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["doctomood"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.messages_control]
disable = [
  "missing-module-docstring",
  "missing-function-docstring",
  "missing-class-docstring",
]
